+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.3.0
+* Added `elevation` parameter.
+* Added `color` parameter.
+* Added `selectedColor` parameter.
+* Added `iconSize` parameter.
+* Added `selectedIconSize` parameter.
+* Added `showLabel` parameter.
+* Added `showSelectedLabel` parameter.
+* Added `backgroundColor` parameter.
+* Removed `padding` parameter.
+* Removed `iconStyle` parameter.
+* Removed `selectedIconStyle` parameter.
+
 ## 0.2.0
 * Added `iconStyle` and `selectedIconStyle` options.
 * Added `padding` option.
@@ -5,3 +18,4 @@
 
 ## 0.1.0
 * Initial Release!
+